refactor: switch small standalone namespaces to self-reexport imports
