+++ packages/opencode/src/effect/app-runtime.ts
@@ -19,7 +19,7 @@ import { Provider } from "@/provider"
 import { ProviderAuth } from "@/provider"
 import { Agent } from "@/agent/agent"
 import { Skill } from "@/skill"
-import { Discovery } from "@/skill"
+import { Discovery } from "@/skill/discovery"
 import { Question } from "@/question"
 import { Permission } from "@/permission"
 import { Todo } from "@/session/todo"
