+++ packages/opencode/src/permission/index.ts
@@ -1,2 +1,1 @@
 export * as Permission from "./permission"
-export * as BashArity from "./arity"
