+++ packages/opencode/src/skill/index.ts
@@ -1,2 +1,1 @@
 export * as Skill from "./skill"
-export * as Discovery from "./discovery"
