+++ packages/opencode/src/tool/bash.ts
@@ -14,7 +14,7 @@ import { fileURLToPath } from "url"
 import { Flag } from "@/flag/flag"
 import { Shell } from "@/shell/shell"
 
-import { BashArity } from "@/permission"
+import { BashArity } from "@/permission/arity"
 import * as Truncate from "./truncate"
 import { Plugin } from "@/plugin"
 import { Effect, Stream } from "effect"
