+++ packages/opencode/test/skill/discovery.test.ts
@@ -1,6 +1,6 @@
 import { describe, test, expect, beforeAll, afterAll } from "bun:test"
 import { Effect } from "effect"
-import { Discovery } from "../../src/skill"
+import { Discovery } from "../../src/skill/discovery"
 import { Global } from "../../src/global"
 import { Filesystem } from "../../src/util"
 import { rm } from "fs/promises"
